HealthComponent handles stun immunity

+++ Assets/Scripts/GameDataManager/GameDataManager.cs
@@ -10,11 +10,11 @@ public static class GameDataManager
         Vector2 position = player.transform.position;
         PlayerPrefs.SetFloat("Player.x", position.x);
         PlayerPrefs.SetFloat("Player.y", position.y);
-        PlayerPrefs.SetFloat("Player.health", player.Health);
+        PlayerPrefs.SetInt("Player.health", player.Health);
     }
     public static void loadData(){
         Player player = GameObject.FindWithTag("Player").GetComponent<Player>();
         player.transform.position = new Vector2(PlayerPrefs.GetFloat("Player.x", player.transform.position.x), PlayerPrefs.GetFloat("Player.y", player.transform.position.y));
-        player.Health = PlayerPrefs.GetFloat("Player.health", 100);
+        player.Health = PlayerPrefs.GetInt("Player.health", 100);
     }
 }
